pyproject.toml
[project]
name = "sqlalchemy-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "faker>=37.0.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.2",
    "sqlalchemy[asyncio]>=2.0.39",
]

[tool.ruff]
src = ["src"]
